Refactored how working set is built and used.
We now use an inernal field when entering new cards and pick from the oldest/smallest ID.  This means cards added will always be picked last.  This fixes a problem where as we added new cards from the pool of non-genesis, those cards are technically older than all the ones in the working set so they always get priority.  Thus, we get stuck in this cycle of constantly churning through a list of easy cards and never goign through the full working set.  This change forces us to always go through the working set as a priority at least once before going into the new cards from the larger pool.