Fixed issues with running on Ubuntu which is production
source wasn't being found, which is because we were using /bin/sh (default) instead of /bin/bash.  It's only defined in bash, so we need to set that variable as defined in the Makefile spec by GNU.  Moreover, we should use double-ampersand && instead of semicolon+newline because the latter will ignore failures from the source command but the former will halt execution as soon as source fails.  Also, make will stop because we aren't ignoring/squashing hte errors from source, which means test and produciton will actually not run if virtualenv fails at source.